% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cv.R
\name{build_cv}
\alias{build_cv}
\title{Build CV}
\usage{
build_cv(
  full_name = "Michael Schramm",
  data_location = NULL,
  peer_review = NULL,
  tech_report = NULL,
  conference = NULL,
  software = NULL,
  datasets = NULL,
  output_file_name = "cv.pdf",
  output_dir = getwd()
)
}
\arguments{
\item{full_name}{My name}

\item{data_location}{character, link to google sheets}

\item{peer_review}{Object of class \code{BibEntry}}

\item{tech_report}{Object of class \code{BibEntry}}

\item{conference}{Object of class \code{BibEntry}}

\item{software}{Object of class \code{BibEntry}}

\item{datasets}{Object of class \code{BibEntry}}

\item{output_file_name}{what should the output doc name be}

\item{output_dir}{output directory, defaults to current working directory}
}
\value{
files written to the current directory
}
\description{
Build CV
}
