% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{create_CV_object}
\alias{create_CV_object}
\title{Create a CV_Printer object.}
\usage{
create_CV_object(
  data_location,
  source_location = "github.com/mps9506",
  sheet_is_publicly_readable = TRUE
)
}
\arguments{
\item{data_location}{Path of the spreadsheets holding all your data. This can be
either a URL to a google sheet with multiple sheets containing the four
data types or a path to a folder containing four \code{.csv}s with the neccesary
data.}

\item{source_location}{Where is the code to build your CV hosted?}

\item{sheet_is_publicly_readable}{If you're using google sheets for data,
is the sheet publicly available? (Makes authorization easier.)}
}
\value{
A new \code{CV_Printer} object.
}
\description{
Create a CV_Printer object.
}
