% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_functions.R
\name{cvR_format_pdf}
\alias{cvR_format_pdf}
\title{cvR format (PDF)}
\usage{
cvR_format_pdf(fig_width = 4, fig_height = 2.5, fig_crop = TRUE,
  dev = "pdf", highlight = "default", keep_tex = FALSE,
  number_sections = FALSE, includes = NULL, md_extensions = NULL,
  pandoc_args = NULL)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures.}

\item{fig_height}{Default height (in inches) for figures.}

\item{fig_crop}{\code{TRUE} to automatically apply the \code{pdfcrop} utility
(if available) to pdf figures.}

\item{dev}{Graphics device to use for figure output (defaults to pdf).}

\item{highlight}{Syntax highlighting style. Default to "pygments", see
\code{\link[rmarkdown]{pdf_document}} for supported styles. Pass \code{NULL}
to prevent syntax hightlighting.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.}

\item{number_sections}{\code{TRUE} to number sections headings.}

\item{includes}{Named list of additional content to include within the
document (typically created using the includes function).}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the
\code{\link[rmarkdown]{rmarkdown_format}} for additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}
}
\description{
Template for creating a CV in PDF format from Rmd file,
  adapted from \code{\link[rmarkdown]{tufte_handout}}
}

