% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cvr_geos_version}
\alias{cvr_geos_version}
\title{GEOS version cvr was built against}
\usage{
cvr_geos_version()
}
\value{
A version string.
}
\description{
GEOS version cvr was built against
}
\examples{
cvr_geos_version()
}
