% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-latexify.R
\name{latexify}
\alias{latexify}
\title{Pipe operator}
\usage{
latexify(x, doublebackslash = TRUE, dashdash = TRUE, quotes = c("straight", "curved"), packages = c("fontenc", "textcomp"))
}
\description{
See \code{dplR::\link[dplR:latexify]{latexify}} for details.
}
\keyword{internal}
