% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.filter.R
\name{my.filter}
\alias{my.filter}
\title{my.filter}
\usage{
my.filter(orcid.input, years.since = 2000, max.authors = 3,
  spacer = ", ", journal.only = "Yes", order = "ayear",
  bold.author = T)
}
\arguments{
\item{orcid.input}{Result of a call to my.orcid() from this package}

\item{years.since}{Yeats of publications you want returned}

\item{max.authors}{Number of authors you want listed}

\item{spacer}{Commas between authors}

\item{journal.only}{Include just journal articles in output. Options 'Yes' or 'No'}

\item{order}{Either 'ayear' for most recent last or 'dyear' for most recent first}

\item{bold.author}{Option to bold author of CV in publications (TRUE or FALSE)}
}
\description{
my.filter
}
