% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctions.R
\name{plot.CVR}
\alias{plot.CVR}
\title{Plot a CVR object.}
\usage{
\method{plot}{CVR}(x, ...)
}
\arguments{
\item{x}{A CVR object.}

\item{...}{Other graphical parameters used in plot.}
}
\description{
Plot the tuning of CVR
}
\details{
The first plot is mean cv error vs log(\eqn{\lambda}). The type of mean cv error is
          decided by \code{type.measure} (see parameters of \code{CVR}). The selected  \eqn{\lambda}  is marked 
          by a vertical line in the plot. The second plot is sparsity vs log(\eqn{\lambda}). 
          Sparsity is the proportion of non-zero elements in fitted W1 and W2. 
          The threshold is marked by a horizontal line. 
          Press ENTER to see the second plot, which shows the tuning of \eqn{\eta}.
}
