% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.papers.R
\name{print.papers}
\alias{print.papers}
\title{Print the publications in cv}
\usage{
\method{print}{papers}(papers)
}
\arguments{
\item{papers}{Result of a call to my.filter()}
}
\description{
Print the publications in cv
}
