% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{print_contact_info}
\alias{print_contact_info}
\title{Print Contact Info}
\usage{
print_contact_info(cv)
}
\arguments{
\item{cv}{cv object}
}
\value{
formatted string
}
\description{
Contact information section with icons
}
\keyword{internal}
