% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_papers.R
\name{print_papers}
\alias{print_papers}
\title{Print the publications in cv}
\usage{
print_papers(papers)
}
\arguments{
\item{papers}{Result of a call to my.filter()}
}
\description{
Print the publications in cv
}
