% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{print_section}
\alias{print_section}
\title{Print Section}
\usage{
print_section(cv, section_id, glue_template = "default", latexify = FALSE)
}
\arguments{
\item{cv}{output from create_cv_object}

\item{section_id}{ID of the entries section to be printed as encoded by the \code{section} column of the \code{entries} table}

\item{glue_template}{glue string}

\item{latexify}{logical, convert strings to latex? default = \code{FALSE}}
}
\value{
formatted string
}
\description{
Take a position data frame and the section id desired and prints the section to markdown.
}
\keyword{internal}
