% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{print_skills_block}
\alias{print_skills_block}
\title{Print Skills Block}
\usage{
print_skills_block(cv)
}
\arguments{
\item{cv}{cv object}
}
\value{
formatted string
}
\description{
Print Skills Block
}
\keyword{internal}
