% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{print_text_block}
\alias{print_text_block}
\title{Print Text Block}
\usage{
print_text_block(cv, label)
}
\arguments{
\item{label}{ID of the text block to print as encoded in \code{label} column of \code{text_blocks} table.}
}
\value{
markdown formatted string
}
\description{
Prints out text block identified by a given label.
}
\keyword{internal}
