% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cv_object.R
\name{sanitize_links}
\alias{sanitize_links}
\title{Sanitize Links}
\usage{
sanitize_links(cv, text)
}
\arguments{
\item{cv}{cv object}

\item{text}{string}
}
\value{
list
}
\description{
Remove links from a text block and add to internal list
}
\keyword{internal}
