% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cv.R
\name{build_cv}
\alias{build_cv}
\title{Build CV}
\usage{
build_cv(
  full_name = "Michael Schramm",
  data_location = NULL,
  peer_review = NULL,
  tech_report = NULL,
  conference = NULL,
  software = NULL,
  datasets = NULL,
  rmd_template = c("svm_cv", "yaac"),
  output_file_name = "cv.pdf",
  output_dir = getwd(),
  spell_check = FALSE,
  yml = list(title = "CV", author = "Michael Schramm", firstname = "Michael", lastname
    = "Schramm", github = "mps9506", phone = "910-232-3760", address = "Bryan, Texas",
    email = "mpschramm@gmail.com", tagline =
    "Researcher | Watersheds, water quality and open science")
)
}
\arguments{
\item{full_name}{character string, my name.}

\item{data_location}{character string, link to google sheets document.}

\item{peer_review}{Object of class \code{BibEntry}.}

\item{tech_report}{Object of class \code{BibEntry}.}

\item{conference}{Object of class \code{BibEntry}.}

\item{software}{Object of class \code{BibEntry}.}

\item{datasets}{Object of class \code{BibEntry}.}

\item{rmd_template}{character string indicating path to rmarkdown template or one of \code{c('svm_cv', 'yaac')}.}

\item{output_file_name}{character string for the name of the output document.}

\item{output_dir}{character string for the name of the output directory, defaults to current working directory.}

\item{spell_check}{logical use the spell_check package to spell check output? defaults to FALSE}

\item{yml}{named list of some yml parameters to pass to the rmd. There is probably a better way to do this.}
}
\value{
files written to the current directory
}
\description{
Build CV
}
