\name{CVR-package}
\alias{CVR-package}
\docType{package}
\title{
\packageTitle{CVR}
}
\description{
\packageDescription{CVR}
}
\details{

\packageIndices{CVR}
functions: \code{cvrsolver}, \code{SparseCCA}, \code{SimulateCVR}, \code{CVR}, \code{plot.CVR}.}
\author{
\packageAuthor{CVR}

Maintainer: \packageMaintainer{CVR}
}
\references{
Chongliang Luo, Jin Liu, Dipak D. Dey and Kun Chen (2016) Canonical variate regression. Accepted by Biostatistics, doi: 10.1093/biostatistics/kxw001.
      
Daniela M. Witten,  Robert Tibshirani and Trevor Hastie (2009) A penalized matrix decomposition, with applications to sparse principal components and canonical correlation analysis. Biostatistics 10(3), 515-534. 

}

\keyword{ package }
\seealso{
PMA.
}

